+++ server/paper/src/main/kotlin/net/stellarica/server/material/custom/item/CustomItemHandler.kt
@@ -1,21 +1,13 @@
 package net.stellarica.server.material.custom.item
 
 import net.minecraft.resources.ResourceLocation
-import net.stellarica.server.StellaricaServer.Companion.klogger
-import net.stellarica.server.StellaricaServer.Companion.plugin
 import net.stellarica.server.material.custom.item.CustomItemHandler.items
-import net.stellarica.server.utils.extensions.customItem
-import net.stellarica.server.utils.extensions.id
-import org.bukkit.Bukkit
-import org.bukkit.Material
-import org.bukkit.NamespacedKey
+import net.stellarica.server.material.type.item.CustomItemType
+import net.stellarica.server.material.type.item.ItemType
 import org.bukkit.event.EventHandler
 import org.bukkit.event.enchantment.PrepareItemEnchantEvent
 import org.bukkit.event.inventory.PrepareAnvilEvent
-import org.bukkit.inventory.ItemStack
-import org.bukkit.inventory.ShapedRecipe
-import org.bukkit.inventory.ShapelessRecipe
-import java.net.http.WebSocket.Listener
+import org.bukkit.event.Listener
 
 /**
  * Keeps track of registed [items]
@@ -30,10 +22,10 @@ object CustomItemHandler: Listener {
 	 */
 	@EventHandler
 	fun onTableEnchant(event: PrepareItemEnchantEvent) {
-		val item = event.item.customItem ?: return
+		val item = ItemType.of(event.item) as? CustomItemType ?: return
 		for (i in 0..2) {
-			val offer = event.offers.get(i) ?: continue
-			if (offer.enchantment !in item.allowedEnchants) event.offers.set(i, null)
+			val offer = event.offers[i] ?: continue
+			if (offer.enchantment !in (item.item.allowedEnchants ?: mutableSetOf())) event.offers[i] = null
 		}
 	}
 
@@ -43,12 +35,10 @@ object CustomItemHandler: Listener {
 	 */
 	@EventHandler
 	fun onAnvilEnchant(event: PrepareAnvilEvent) {
-		val item = event.result?.customItem ?: return
+		val item = event.result?.let { ItemType.of(it) } as? CustomItemType ?: return
 		event.result!!.enchantments.keys.forEach {
-			if (it !in item.allowedEnchants) event.result = null
+			if (it !in (item.item.allowedEnchants ?: mutableSetOf())) event.result = null
 			return
 		}
 	}
-
-
 }
