--- server/paper/src/main/kotlin/net/stellarica/server/utils/ChunkLocation.kt
@@ -1,7 +1,0 @@
-package net.stellarica.server.utils
-
-/**
- * Container for an [x] and [z] of a chunk
- * @see BlockPos
- */
-data class ChunkLocation(var x: Int, var z: Int)
